Pulled out the commonality into a base class
